refactor(Validator): 자동차 이름, 시행횟수를 검증하는 부분을 Validator 로 별도 추출
